+++ max-temp/clean.sh
@@ -1,8 +1,17 @@
 #!/bin/bash
 
+while getopts "f" opt; do
+  case $opt in
+    f) force=true ;;
+  esac
+done
+
 REPO_TOP=$(git rev-parse --show-toplevel)
 input_dir="${REPO_TOP}/max-temp/inputs"
 outputs_dir="${REPO_TOP}/max-temp/outputs"
 
-rm -rf "${input_dir}"
 rm -rf "${outputs_dir}"
+
+if [ "$force" = true ]; then
+    rm -rf "${input_dir}"
+fi
