+++ media-conv/clean.sh
@@ -1,8 +1,17 @@
 #!/bin/bash
 
+while getopts "f" opt; do
+  case $opt in
+    f) force=true ;;
+  esac
+done
+
 REPO_TOP=$(git rev-parse --show-toplevel)
 outputs_dir="${REPO_TOP}/media-conv/outputs"
 input_dir="${REPO_TOP}/media-conv/inputs"
 
 rm -rf "$outputs_dir"
-rm -rf "$input_dir"
+
+if [ "$force" = true ]; then
+    rm -rf "$input_dir"
+fi
