+++ nlp/clean.sh
@@ -1,6 +1,15 @@
 #!/bin/bash
 
+while getopts "f" opt; do
+  case $opt in
+    f) force=true ;;
+  esac
+done
+
 cd "$(realpath "$(dirname "$0")")" || exit 1
 
-rm -rf ./inputs
 rm -rf ./outputs
+
+if [ "$force" = true ]; then
+    rm -rf ./inputs
+fi
